+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Allow tool result to be the last message in `AnthropicTextProvider`.
+
 ## [0.5.0] - 2024-08-19
 
 ### Changed